requirements.txt
PyYAML
distro
tqdm
termcolor